webdriver: Make max threads configurable
This PR introduces a new permanent feature flag called
max-selenium-threads which allows us to specify the number of concurrent
tests to run when executing the selenium simulator.

This also bumps the version of Selenium to a more recent release which
fixed a bug where multiple versions of chrome were being launched.