webdriver: Run simulator for specific environments
This commit allows us to add the `nosim` tag to an environment in
LaunchDarkly which will cause the Simulator not to run against this
environment.

The purpose here is to only run the Simulator for environments where
people may actually be showing a demo.

Otherwise we are wasting both CI resources and cogs by having to index
many fake users whose data will never be used.