webdriver: Allow disabeling the simulator
This commit allows us to use a new feature flag called
`disable-simulator` which allows us to disable the simulator. This is
useful in case we need to debug some things without the additional noise
of the simulator.

Right now the simulator runs every 15 minutes so the logs that we see
are full of tons of information that is very noisy.